+++ projects/iad-interface-admin/form/src/public_api.ts
@@ -1,11 +1,43 @@
+// Dynamic form components
+export * from './lib/dynamic-form/components/iad-dropdown-group.component';
+export * from './lib/dynamic-form/components/iad-form-checkbox.component';
+export * from './lib/dynamic-form/components/iad-form-chips.component';
+export * from './lib/dynamic-form/components/iad-form-date.component';
+export * from './lib/dynamic-form/components/iad-form-date-time.component';
+export * from './lib/dynamic-form/components/iad-form-file.component';
+export * from './lib/dynamic-form/components/iad-form-input.component';
+export * from './lib/dynamic-form/components/iad-form-multi-select.component';
+export * from './lib/dynamic-form/components/iad-form-number.component';
+export * from './lib/dynamic-form/components/iad-form-rich-editor.component';
+export * from './lib/dynamic-form/components/iad-form-selection-dropdown.component';
+export * from './lib/dynamic-form/components/iad-form-textarea.component';
+export * from './lib/dynamic-form/components/iad-form-translate-input.component';
+
+// Dynamic form models
+export * from './lib/dynamic-form/inputs/auto-complete-input.model';
+export * from './lib/dynamic-form/inputs/boolean-input.model';
+export * from './lib/dynamic-form/inputs/chips-input.model';
+export * from './lib/dynamic-form/inputs/date-input.model';
+export * from './lib/dynamic-form/inputs/date-time-input.model';
+export * from './lib/dynamic-form/inputs/dropdown-input.model';
+export * from './lib/dynamic-form/inputs/file-input.model';
+export * from './lib/dynamic-form/inputs/hidden-input.model';
+export * from './lib/dynamic-form/inputs/multi-select-input.model';
+export * from './lib/dynamic-form/inputs/number-input.model';
+export * from './lib/dynamic-form/inputs/rich-editor-input.model';
+export * from './lib/dynamic-form/inputs/text-input.model';
+export * from './lib/dynamic-form/inputs/text-input-translated.model';
+export * from './lib/dynamic-form/inputs/textarea-input.model';
+
 // Dynamic form
 export * from './lib/dynamic-form/core/context-aware';
 export * from './lib/dynamic-form/core/validation-input';
 export * from './lib/dynamic-form/core/form-input-group';
 export * from './lib/dynamic-form/core/input.factory';
 export * from './lib/dynamic-form/core/form-input.model';
-export * from './lib/dynamic-form/dynamic-form.module';
+export * from './lib/dynamic-form/dynamic-form.component';
 export * from './lib/dynamic-form/services/iad-field-values.service';
+export * from './lib/dynamic-form/dynamic-form.module';
 
 // projection form
 export * from './lib/projection-form/model/iad-form-projection.model';