+++ Changelog.md
@@ -1,4 +1,4 @@
-# 1.0.7@alpha.86
+# 1.0.7@alpha.88
 
 * GRID: Added ability to set [Angular Perfect Scrollbar](https://github.com/zefoy/ngx-perfect-scrollbar) for grid
 
@@ -10,6 +10,28 @@
     ```
     
     NOTE: It will make any sense only if enableInfiniteScroll is true and grid has absolute positioning with set up width and height;
+
+# 1.0.7-alpha.78
+
+* Form disableFooter changed to isNestedForm. It will not only disable form footer, but also disable the whole <form> tag, leaving clear <fieldset> inner content 
+
+# 1.0.7-alpha.77
+
+* Form Export of DynamicFormComponent
+* Form disableFooter option will disable form footr when set to true
+* Form added outgoing valueChanges event
+
+
+# 1.0.7-alpha.75
+
+* FORM added support for selection-restricted values for autocomplete
+* GRID changed minimum input length for delete action in action column
+
+# 1.0.7-alpha.74
+
+* FORM added export of all dynamic form models and components
+* FORM added dropdown options valueField: string; and  labelField: string;
+* FORM added date and datetime option dateFormat: string; to set displaying date
 
 # 1.0.7-alpha.73